Update and rename endPoints.js to index.js
